bat_migration_europe/requirements.txt

# YAML:
pyyaml
#dict2xml
# Timezones:
pytz
# Command line:
click
# VS Code:
pylint
black
